FIX - Screen-QR 폴더 작업하기 쉽게 분리
QR화면에서 -
MaterialTopTabNavigation을 타고 -
Keyboard & QRScanner를 띄워서 -
거기서 InputNumber를  Navigation으로 타서-
NumberCheck, NumberComp을 Stack Screen으로 띄우는
과정을 파악하기 쉽게 수정했습니다
