+++ App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import BotNav from './navigation/bottomNav';
+import BotNav from './navigation/App_bottomNav';
 
 export default function App() {
 	return (
