+++ screens/inputScreens/stackScreen/NumberComp.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 
-export default function log() {
+export default function Comp() {
 	return (
 		<View
 			style={{
